Allow opening activity to specify orientation.
An activity's window is initially not visible. This means it is not
accounted for until it becomes visible, which means requested
orientations will not be honored until after onCreate.

This changelist removes the visibility check previously used determining
the orientation.

This changelist also addresses an issue where an ActivityRecord member
variable tracking the last configuration sent was not being initially
updated.

Fixes: 33844887
Fixes: 33814250
Fixes: 34165818
Fixes: 34177026
Fixes: 33809086
Fixes: 34111451

Test: Verified reported issues fixed in apps mentioned in bugs
Test: Added additional unit tests to exercise override getOrientation.

Change-Id: I26b7e808b6f3d857543e3a187b6c318e88afa063
Signed-off-by: Bryce Lee <brycelee@google.com>