Fix issue #37306208: CTS: ServiceTest#testForegroundService...
..._removeNotificationOnStop failure

We weren't removing the foreground timeout when a service is
brought down and still waiting for the app to make it foreground.
Instead, turn this in to an immediate crash of the app when we
detect it while bringing down the service.  Also make various
paths more robust against trying to do things with a ServiceRecord
that isn't currently in ActiveServices.

Test: bit CtsAppTestCases:ServiceTest

Change-Id: I06f4ac48fb983483569677ab1ac6e5b0b681c4ae