Skip Pip animation to fullscreen if orientation is going to change
Having this kind of animation look good is extremely difficult.
The best we can do is skip the animation and just go to fullscreen
if we know the device orientation is going to change because the
current screen configuration isn't compatible with the fixed
orientation requested by the app going fullscreen.

Change-Id: I97b14723a0d678c05efedc93fe692ad3b3212a72
Fixes: 37722472
Test: manual