Fixed CTS failure due to selecting wrong transition.
ag/2253646 fixed an issue where a starting window can be displayed prematurely.
However, it also prevented an app from being added to the opening apps list if
it was already visible. This caused wrong transition animation to be selected
for the closing app. We now always set the visibility to true even if the app
is already visible so that the app can be added to the opening apps list and
the right transition animation selected. We don't set visibility to false if
the app isn't visible so it isn't added to the closing apps list which covers
the case ag/2253646 attempted to fix.

Bug: 34545029
Change-Id: I06363beb0c524146a8b871eff8157678452c3d1d
Fixes: 38505794
Test: cts/.../run-test android.server.cts.ActivityManagerTransitionSelectionTests