Add modified API to send more information to VrCore
Currently, null is sent when a 2D intent is running.
Sending the component name and a flag instead allows
VrCore to determine which intent is currently running
and which layers on the display correspond to this
application.

The process ID of the current intent is also sent,
to allow association of buffers to the VR app.

Bug: 63709047, 63115025
Test: manual with prints

Change-Id: I164b577f2c578867fb953d92074ef3d2d31221a1