Make sure top activity is stopped on sleep if paused.
Currently we have logic in place to handle moving a top resumed
activity to paused and eventually stopped if entering sleep. However,
we do not handle the case when the top activity is already paused.

This changelist addresses this issue by identifying this condition
and moving the activity to the stopped state.

Change-Id: Ib9fff1efeb33719d008508cf21a17b95fc17efd0
Fixes: 65457314
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerDisplayTests#testExternalDisplayActivityTurnPrimaryOff
(cherry picked from commit 1e7c437b58ca26f81a3192ae3ab857562094bbd6)