Allow modifiers to specify windowing mode and preferred display.
This changelist extends the parameters that positioners can
provide to include the windowing mode and display for a stack to
launch into. These are provided via the LaunchingParams object, which
replaces the previous bounds Rect as the result. When the values are
provided when laying out a task, the Controller will apply the
result. In other cases, such as in ActivityStarter, the caller of
LaunchingBoundsController#calculate is responsible applying the
result.

Fixes: 70031444
Test: atest FrameworksServicesTests:LaunchingBoundsControllerTests#testLayoutTaskPreferredDisplayChange
Test: atest FrameworksServicesTests:LaunchingBoundsControllerTests#testLayoutTaskWindowingModeChange
Change-Id: I949e2b5047663bfc297aac654f365109f0b1feb1